CORE: Membership expiration notification logic moved to registrar
- Part of Synchronizer, where we check membership status based on
  membership expiration attribute was moved to perun-registrar-lib
  project in ExpirationNotifScheduler.
  Reason is to add ability to skip notification about expiration,
  when user has submitted (pending) extension application in same VO.
- Updated spring context - removed method calls from
  perun-core-synchronizers.xml and added to
  perun-registrar-lib-scheduler.xml.
- Tests of Synchronizer removed and logic moved to
  ExpirationNotifSchedulerTest.
- Use new context file in perun-rpc.
