+++ perun-registrar-lib/src/main/resources/perun-registrar-lib-scheduler.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xmlns:task="http://www.springframework.org/schema/task"
+       xsi:schemaLocation="
+http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
+http://www.springframework.org/schema/task http://www.springframework.org/schema/task/spring-task.xsd
+">
+
+	<task:scheduler id="scheduler" pool-size="1"/>
+
+	<beans profile="production">
+		<task:scheduled-tasks scheduler="scheduler">
+			<task:scheduled ref="registrarScheduler" method="checkMembersState" cron="0 5 0 * * ?"/> <!-- every day at 00:05 -->
+		</task:scheduled-tasks>
+	</beans>
+
+	<beans profile="devel">
+		<task:scheduled-tasks scheduler="scheduler">
+			<task:scheduled ref="registrarScheduler" method="checkMembersState" cron="0 5 0 * * ?"/> <!-- every day at 00:05 -->
+		</task:scheduled-tasks>
+	</beans>
+
+</beans>
