+++ perun-registrar-lib/src/main/resources/perun-registrar-lib.xml
@@ -41,6 +41,11 @@ http://www.springframework.org/schema/tx http://www.springframework.org/schema/t
 		<property name="registrarManager" ref="registrarManager"/>
 	</bean>
 
+	<bean id="registrarScheduler" class="cz.metacentrum.perun.registrar.impl.ExpirationNotifScheduler" init-method="initialize">
+		<property name="dataSource" ref="dataSource"/>
+		<property name="perun" ref="perun"/>
+	</bean>
+
 	<!-- Properties Bean -->
 	<beans profile="devel,production">
 		<bean id="registrarProperties" class="org.springframework.beans.factory.config.PropertiesFactoryBean">
